// src/api-client/models/address-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EntityBasic } from './entity-basic';
// May contain unused imports in some cases
// @ts-ignore
import type { WardBasic } from './ward-basic';

/**
 * 
 * @export
 * @interface AddressResponse
 */
export interface AddressResponse {
    /**
     * 
     * @type {number}
     * @memberof AddressResponse
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof AddressResponse
     */
    'isDefault'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AddressResponse
     */
    'isVisible'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof AddressResponse
     */
    'status'?: AddressResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof AddressResponse
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressResponse
     */
    'streetDetail'?: string;
    /**
     * 
     * @type {WardBasic}
     * @memberof AddressResponse
     */
    'ward'?: WardBasic;
    /**
     * 
     * @type {EntityBasic}
     * @memberof AddressResponse
     */
    'district'?: EntityBasic;
    /**
     * 
     * @type {EntityBasic}
     * @memberof AddressResponse
     */
    'province'?: EntityBasic;
}

export const AddressResponseStatusEnum = {
    Active: 'ACTIVE',
    Inactive: 'INACTIVE'
} as const;

export type AddressResponseStatusEnum = typeof AddressResponseStatusEnum[keyof typeof AddressResponseStatusEnum];